# cmd/docs/swagger.yaml
definitions:
...
  aqary_admin_internal_domain_requests_user.CreateRoleRequest:
    properties:
      department_id:
        type: integer
      permissions:
        type: string
      role:
        type: string
    required:
    - department_id
    - role
    type: object
...
  aqary_admin_internal_domain_requests_user.UpdateRoleRequest:
    properties:
      department_id:
        type: integer
      permissions:
        type: string
      role:
        type: string
    required:
    - department_id
    type: object
...
  aqary_admin_internal_domain_requests_user.UserTypeRequest:
    properties:
      user_type:
        type: string
    type: object
...
  aqary_admin_internal_domain_responses_user.AllRolesOutput:
    properties:
      id:
        type: integer
      is_role_exist:
        type: boolean
      label:
        type: string
    type: object
...
  aqary_admin_internal_domain_sqlc_sqlc.Role:
    properties:
      created_at:
        type: string
      department_id:
        $ref: '#/definitions/pgtype.Int8'
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  aqary_admin_internal_domain_sqlc_sqlc.UserType:
    properties:
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_type:
        type: string
    type: object
...
  aqary_admin_pkg_utils.ErrResponseSwagger:
    properties:
      error:
        example: any error message
        type: string
    type: object
...
  pgtype.Int8:
    properties:
      int64:
        type: integer
      valid:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Testing Swagger APIs.
  title: Aqary  APIs
  version: "1.0"
paths:
...
  /api/user/createRole:
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Create a new role
      tags:
      - roles
...
  /api/user/createusertype:
    post:
      description: CreateUserType
      parameters:
      - description: UserTypeRequest
        in: body
        name: UserTypeRequest
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.UserTypeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.UserType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: CreateUserType
      tags:
      - usertype
...
  /api/user/deleteRole/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Delete a role
      tags:
      - roles
...
  /api/user/deleteusertype/:id:
    delete:
      description: DeleteUserType
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.UserType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: DeleteUserType
      tags:
      - usertype
...
  /api/user/getAllRoles:
    get:
      consumes:
      - application/json
      description: Get all roles with pagination
      parameters:
      - description: Page size
        in: query
        name: page_size
        required: true
        type: integer
      - description: Page number
        in: query
        name: page_no
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all roles with pagination
      tags:
      - roles
  /api/user/getAllRolesWithoutPagination/all:
    get:
      consumes:
      - application/json
      description: Get all roles without pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_responses_user.AllRolesOutput'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all roles without pagination
      tags:
      - roles
...
  /api/user/getAlluserTypeForWeb:
    get:
      description: GetAllUserTypesForWeb
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.UserType'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: GetAllUserTypesForWeb
      tags:
      - usertype
...
  /api/user/getRole/{id}:
    get:
      consumes:
      - application/json
      description: Get a role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get a role by ID
      tags:
      - roles
...
  /api/user/getallusertype:
    get:
      description: GetAllUserTypes
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.UserType'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: GetAllUserTypes
      tags:
      - usertype
  /api/user/getusertype/:id:
    get:
      description: CreateUserType
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.UserType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: CreateUserType
      tags:
      - usertype
...
  /api/user/updateRole/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Update a role
      tags:
      - roles
...
  /api/user/updateusertype/:id:
    put:
      description: UpdateUserType
      parameters:
      - description: UserTypeRequest
        in: body
        name: UserTypeRequest
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.UserTypeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.UserType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: UpdateUserType
      tags:
      - usertype
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"